pyproject.toml
[project]
name = "llm-starter"
version = "0.1.0"
description = "A scaffold to make a LLM App."
authors = [{ name = "jtsang4", email = "info@jtsang.me" }]
dependencies = [
    "langchain-openai>=0.1.3",
    "python-dotenv>=1.0.1",
    "langfuse>=2.26.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
